+++ docs/conf.py
@@ -73,7 +73,9 @@ extensions = [
     "sphinx.ext.mathjax",
     "sphinx.ext.napoleon",
     "sphinx.ext.githubpages",
+    "myst_nb"
 ]
+# pip install myst-nb
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
@@ -89,7 +91,7 @@ master_doc = "index"
 
 # General information about the project.
 project = "dimcat"
-copyright = "2022, Digital and Cognitive Musicology Lab @ École Polytechnique Fédérale de Lausanne"
+copyright = "2023, Digital and Cognitive Musicology Lab @ École Polytechnique Fédérale de Lausanne"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -247,7 +249,7 @@ latex_elements = {
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, documentclass [howto/manual]).
 latex_documents = [
-    ("index", "user_guide.tex", "dimcat Documentation", "johentsch", "manual")
+    ("index", "user_guide.tex", "dimcat Documentation", "johentsch", "Digital and Cognitive Musicology Lab @ École Polytechnique Fédérale de Lausanne", "manual")
 ]
 
 # The name of an image file (relative to this directory) to place at the top of