+++ src/dimcat/__init__.py
@@ -1,10 +1,4 @@
-import sys
-
-if sys.version_info[:2] >= (3, 8):
-    # TODO: Import directly (no need for conditional) when `python_requires = >= 3.8`
-    from importlib.metadata import PackageNotFoundError, version  # pragma: no cover
-else:
-    from importlib_metadata import PackageNotFoundError, version  # pragma: no cover
+from importlib_metadata import PackageNotFoundError, version  # pragma: no cover
 
 try:
     # Change here if project is renamed and does not equal the package name
@@ -18,8 +12,21 @@ finally:
 from .analyzer import (
     ChordSymbolBigrams,
     ChordSymbolUnigrams,
+    LocalKeySequence,
+    LocalKeyUnique,
     PitchClassVectors,
     TPCrange,
 )
-from .data import Corpus
-from .writer import TSVwriter
+from .data import Dataset
+from .filter import HasCadenceAnnotationsFilter, IsAnnotatedFilter
+from .grouper import CorpusGrouper, ModeGrouper, PieceGrouper, YearGrouper
+from .pipeline import Pipeline
+from .slicer import (
+    ChordCriterionSlicer,
+    ChordFeatureSlicer,
+    LocalKeySlicer,
+    MeasureSlicer,
+    NoteSlicer,
+    PhraseSlicer,
+)
+from .writer import TSVWriter