+++ src/dimcat/pipeline.py
@@ -1,6 +1,6 @@
 """A Pipeline performs computation iteratively on a List of PipelineSteps."""
 from abc import ABC, abstractmethod
-from typing import List
+from typing import List, Tuple
 
 from .data import Data
 
@@ -17,6 +17,21 @@ class PipelineStep(ABC):
         self.required_facets = []
         """Specifies a list of facets (such as 'notes' or 'labels') that the passed Data object
         needs to provide."""
+
+    def check(self, _) -> Tuple[bool, str]:
+        """Test piece of data for certain properties before computing analysis.
+
+        Returns
+        -------
+        :obj:`bool`
+            True if the passed data is eligible.
+        :obj:`str`
+            Error message in case the passed data is not eligible.
+        """
+        return True, ""
+
+    def filename_factory(self):
+        return self.__class__.__name__
 
     @abstractmethod
     def process_data(self, data: Data) -> Data: